Added intent and multi-file functionality
Multi-file functionality has been added. It's slow and kind of buggy
but it seems to work in my experience. The usage has not changed for this.
Also added an optional argument for intent, where the user can pass in what
they think the program should do, and this in turn will be passed into the prompt
sent to GPT-4.